ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.0
python-dotenv==1.0.1
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0
opencv-python==4.8.1.78
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
pyyaml==6.0.1
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
